requirements.txt
streamlit>=1.30.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
pandas>=2.0.0
Pillow>=10.0.0
easyocr>=1.7.0
numpy>=1.24.0
